requirements.txt
# Flask web framework
Flask==3.0.0
Werkzeug==3.0.1

# OpenAI and Azure OpenAI for DIAL API integration
openai==1.55.3

# Environment variables management
python-dotenv==1.0.1

# HTTP requests for API calls
requests==2.32.3

# Image processing
Pillow==10.4.0

# Additional Flask extensions (optional but useful)
Flask-CORS==4.0.0

# Date and time handling
python-dateutil==2.8.2

# JSON handling (usually included with Python, but ensuring compatibility)
certifi==2023.11.17

# For handling file uploads and forms
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.7.0